+++ src/python/PythonSDK/foundationallm/exceptions/__init__.py
@@ -1,0 +1,1 @@
+from .foundationallm_exception import FoundationaLLMException
