+++ src/python/PythonSDK/foundationallm/models/operations/__init__.py
@@ -1,3 +1,4 @@
 from .long_running_operation_log_entry import LongRunningOperationLogEntry
 from .long_running_operation import LongRunningOperation
 from .operation_status import OperationStatus
+from .operation_types import OperationTypes
