+++ src/python/PythonSDK/foundationallm/models/operations/operation_types.py
