+++ src/python/PythonSDK/foundationallm/models/orchestration/__init__.py
@@ -1,13 +1,14 @@
 """FoundationaLLM orchestration models module"""
-from .message_history_item import MessageHistoryItem
+from .analysis_result import AnalysisResult
 from .citation import Citation
-from .completion_request_object_keys import CompletionRequestObjectKeys
-from .completion_request_base import CompletionRequestBase
-from .completion_response import CompletionResponse
-from .operation_types import OperationTypes
+
 from .message_content_item_types import MessageContentItemTypes
 from .message_content_item_base import MessageContentItemBase
+
 from .openai_file_path_message_content_item import OpenAIFilePathMessageContentItem
 from .openai_image_file_message_content_item import OpenAIImageFileMessageContentItem
 from .openai_text_message_content_item import OpenAITextMessageContentItem
-from .analysis_result import AnalysisResult
+
+from .completion_request_object_keys import CompletionRequestObjectKeys
+from .completion_request_base import CompletionRequestBase
+from .completion_response import CompletionResponse
